requirements.txt
asn1crypto==1.5.1
bit==0.7.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
coincurve==18.0.0
dominate==2.9.1
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.10
protobuf==4.21.12
pycparser==2.21
python-dotenv==1.0.0
requests==2.26.0
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.7
